refatoracao para persistir vendas

+++ src/main/resources/db/migration/V2__create_product_table.sql
@@ -5,6 +5,6 @@ create table if not exists products.product(
     product_identifier varchar(20) not null,
     nome varchar(60) not null,
     descricao varchar(120) not null,
-    preco float not null,
+    preco numeric(15,2) not null,
     category_id bigint not null REFERENCES products.category(id)
 );
