+++ src/main/resources/db/migration/V5__create_shop_table.sql
@@ -4,7 +4,7 @@ create table if not exists shopping.shop (
     
     id bigserial primary key,
     data timestamp not null,
-    total float not null,
+    total numeric(15,2) not null,
     user_id bigint not null REFERENCES users.user(id)
 
 );
@@ -12,7 +12,7 @@ create table if not exists shopping.shop (
 create table if not exists shopping.itens (
     id bigserial primary key,
     quantidade int not null,
-    preco float not null,
+    preco numeric(15,2) not null,
     shop_id bigint not null REFERENCES shopping.shop(id),
     product_id bigint not null REFERENCES products.product(id)
 );
